refactor : 이벤트핸들러 생성 및 MockMvc 활용
